organizando components de compra e venda
ajustando nome de itens para condizer mais com sua
usabilidade
